requirements.txt
bip39
requests
py-bip39-bindings
pynacl